Stabilize `let_chains`
