+++ src/test/ui/expr/if/attrs/let-chains-attr.rs
@@ -1,7 +1,5 @@
 // check-pass
 
-#![feature(let_chains)]
-
 #[cfg(FALSE)]
 fn foo() {
     #[attr]
