+++ src/test/ui/mir/mir_let_chains_drop_order.rs
@@ -4,7 +4,6 @@
 
 // See `mir_drop_order.rs` for more information
 
-#![feature(let_chains)]
 #![allow(irrefutable_let_patterns)]
 
 use std::cell::RefCell;
