+++ src/test/ui/rfc-2497-if-let-chains/disallowed-positions.rs
@@ -17,8 +17,6 @@
 //
 // To that end, we check some positions which is not part of the language above.
 
-#![feature(let_chains)] // Avoid inflating `.stderr` with overzealous gates in this test.
-
 #![allow(irrefutable_let_patterns)]
 
 use std::ops::Range;
@@ -100,6 +98,12 @@ fn _macros() {
     //~^ ERROR `let` expressions are not supported here
     //~| ERROR `let` expressions are not supported here
     //~| ERROR expected expression, found `let` statement
+    use_expr!(true && let 0 = 1);
+    //~^ ERROR expected expression, found `let` statement
+
+    macro_rules! noop_expr { ($e:expr) => {}; }
+    noop_expr!((let 0 = 1));
+    //~^ ERROR expected expression, found `let` statement
 }
 
 fn nested_within_if_expr() {
@@ -467,4 +471,7 @@ fn with_parenthesis() {
         ([1, 2, 3][let _ = ()])
         //~^ ERROR expected expression, found `let` statement
     }
+
+    #[cfg(FALSE)] (let 0 = 1);
+    //~^ ERROR expected expression, found `let` statement
 }
