+++ src/test/ui/rfc-2497-if-let-chains/invalid-let-in-a-valid-let-context.rs
@@ -1,5 +1,3 @@
-#![feature(let_chains)]
-
 fn main() {
     let _opt = Some(1i32);
