+++ src/test/ui/rfc-2497-if-let-chains/issue-90722.rs
@@ -1,7 +1,5 @@
 // check-pass
 
-#![feature(let_chains)]
-
 fn main() {
     let x = Some(vec!["test"]);
