+++ src/test/ui/rfc-2497-if-let-chains/issue-92145.rs
@@ -1,7 +1,5 @@
 // check-pass
 
-#![feature(let_chains)]
-
 fn main() {
     let opt = Some("foo bar");
