+++ src/tools/clippy/tests/ui/needless_late_init.rs
@@ -1,5 +1,4 @@
 // run-rustfix
-#![feature(let_chains)]
 #![allow(
     unused,
     clippy::assign_op_pattern,
